Rename appMigrationMetadata
